+++ app/src/main/java/com/example/cscb07project/CommentsFragment.java
@@ -12,7 +12,10 @@ import android.widget.EditText;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import android.content.res.ColorStateList;
+import android.graphics.Color;
 
+import com.google.android.material.button.MaterialButton;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
@@ -46,7 +49,9 @@ public class CommentsFragment extends Fragment{
 
     private RecyclerView commentRecyclerView;
     private CommentAdapter commentAdapter;
-
+    private MaterialButton backButton;
+    private MaterialButton deleteCommentsButton;
+    private boolean deleteModeEnabled = false;
 
     public CommentsFragment() {
     }
@@ -73,13 +78,13 @@ public class CommentsFragment extends Fragment{
         }
         commentList = new ArrayList<>();
         commentRecyclerView = view.findViewById(R.id.commentsRecyclerView);
+        deleteCommentsButton = view.findViewById(R.id.buttonDeleteSelectedComments);
+        backButton = view.findViewById(R.id.back_button);
 
-        commentAdapter = new CommentAdapter(commentList,(comment) -> this.deleteComment(comment));
-        commentRecyclerView.setLayoutManager(
-                new LinearLayoutManager(requireContext())
-        );
 
-        commentRecyclerView.setAdapter(commentAdapter);
+
+
+        updateDeleteButton();
         displayALLComments();
         return view;
     }
@@ -87,23 +92,61 @@ public class CommentsFragment extends Fragment{
     @SuppressLint("NotifyDataSetChanged")
     public void displayALLComments(){
 
-        commentAdapter = new CommentAdapter(commentList, comment -> deleteComment(comment));
+        commentAdapter = new CommentAdapter(commentList, comment -> deleteComment(comment),deleteModeEnabled);
 
         commentRecyclerView.setLayoutManager(new LinearLayoutManager(requireContext()));
         commentRecyclerView.setAdapter(commentAdapter);
+        deleteCommentsButton.setOnClickListener(v -> changeDeleteMode());
+        backButton.setOnClickListener(v -> returnToExpandedView());
+
 
         if(currentLotNumber==null){return;}
         expandedViewRepo.getCommentsByLotNumber(currentLotNumber).addOnSuccessListener( comments ->{
-                    commentList.clear();
-                    if(commentList!=null){
-                        commentList.addAll(comments);
-                    }
-                    commentAdapter.notifyDataSetChanged();
-                });
+            commentList.clear();
+            if(comments!=null){
+                commentList.addAll(comments);
+            }
+            commentAdapter.notifyDataSetChanged();
+        });
 
 
     }
+    private void changeDeleteMode() {
+        deleteModeEnabled = !deleteModeEnabled;
+        updateDeleteButton();
+        commentAdapter.setDeleteModeEnabled(deleteModeEnabled);
+    }
+
+    private void updateDeleteButton() {
+       int red = Color.rgb(183, 40, 45);
+        if (deleteModeEnabled) {
+            deleteCommentsButton.setBackgroundTintList(
+                    ColorStateList.valueOf(red)
+            );
+
+            deleteCommentsButton.setTextColor(Color.WHITE);
+
+            deleteCommentsButton.setIconTint(
+                    ColorStateList.valueOf(Color.WHITE)
+            );
+        } else {
+            deleteCommentsButton.setBackgroundTintList(
+                    ColorStateList.valueOf(Color.WHITE)
+            );
+
+            deleteCommentsButton.setTextColor(red);
+
+            deleteCommentsButton.setIconTint(
+                    ColorStateList.valueOf(red)
+            );
+        }
+    }
+    @SuppressLint("NotifyDataSetChanged")
+
     private void deleteComment(Comment comment) {
+        if (!deleteModeEnabled) {
+            return;
+        }
         if (comment ==null||comment.getCommentId()==null) {return;}
 
         expandedViewRepo.deleteCommentById(currentLotNumber, comment.getCommentId())
@@ -122,5 +165,15 @@ public class CommentsFragment extends Fragment{
                                 Toast.LENGTH_SHORT
                         ).show()
                 );
+    }
+
+    private void returnToExpandedView() {
+        if (currentLotNumber == null || currentLotNumber.trim().isEmpty()) {
+            return;
+        }
+        backButton.setEnabled(false);
+        ExpandedArtifactFragment expandedArtifactFragment = ExpandedArtifactFragment.newInstance(currentLotNumber);
+        requireActivity().getSupportFragmentManager().beginTransaction().replace(R.id.fragment_container, expandedArtifactFragment).addToBackStack(null).commit();
+        backButton.setEnabled(true);
     }
 }
